Move SpaceModel to common/
