+++ src/client/components/moon/MoonBoard.vue
@@ -76,7 +76,7 @@
 
 import Vue from 'vue';
 import {MoonModel} from '@/models/MoonModel';
-import {SpaceModel} from '@/models/SpaceModel';
+import {SpaceModel} from '@/common/models/SpaceModel';
 import {SpaceType} from '@/common/boards/SpaceType';
 import MoonSpace from '@/client/components/moon/MoonSpace.vue';
