+++ src/common/models/SpaceModel.ts
@@ -1,9 +1,8 @@
-
-import {Color} from '../common/Color';
-import {TileType} from '../common/TileType';
-import {SpaceBonus} from '../common/boards/SpaceBonus';
-import {SpaceType} from '../common/boards/SpaceType';
-import {SpaceId} from '../common/Types';
+import {Color} from '../Color';
+import {TileType} from '../TileType';
+import {SpaceBonus} from '../boards/SpaceBonus';
+import {SpaceType} from '../boards/SpaceType';
+import {SpaceId} from '../Types';
 
 export type SpaceHighlight = undefined | 'noctis' | 'volcanic';
